+++ packages/web3/src/accounts.ts
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { EthExecutionAPI, Bytes, Transaction } from 'web3-types';
+import { EthExecutionAPI, Bytes, Transaction, KeyStore } from 'web3-types';
 import { ETH_DATA_FORMAT, format } from 'web3-utils';
 import { Web3Context } from 'web3-core';
 import { prepareTransactionForSigning } from 'web3-eth';
@@ -59,7 +59,7 @@ export const initAccountsForContext = (context: Web3Context<EthExecutionAPI>) =>
 	};
 
 	const decryptWithContext = async (
-		keystore: string,
+		keystore: KeyStore | string,
 		password: string,
 		options?: Record<string, unknown>,
 	) => {